+++ projects/angular-cesium/src/lib/angular-cesium-widgets/models/editable-circle.ts
@@ -5,7 +5,8 @@ import { Cartesian3 } from '../../angular-cesium/models/cartesian3';
 import { GeoUtilsService } from '../../angular-cesium/services/geo-utils/geo-utils.service';
 import { EditArc } from './edit-arc';
 import { CircleEditOptions } from './circle-edit-options';
-import { PointProps, PolylineProps } from './polyline-edit-options';
+import { PointProps } from './point-edit-options';
+import { PolylineProps } from './polyline-edit-options';
 import { defaultLabelProps, LabelProps } from './label-props';
 import { EllipseProps } from './ellipse-edit-options';