+++ projects/angular-cesium/src/lib/angular-cesium-widgets/models/editor-edit-options.ts
@@ -1,6 +1,7 @@
 import { CesiumEvent } from '../../angular-cesium/services/map-events-mananger/consts/cesium-event.enum';
 import { CesiumEventModifier } from '../../angular-cesium/services/map-events-mananger/consts/cesium-event-modifier.enum';
-import { PointProps, PolylineProps } from './polyline-edit-options';
+import { PointProps } from './point-edit-options';
+import { PolylineProps } from './polyline-edit-options';
 import { PickConfiguration } from '../../angular-cesium';
 
 export interface EditorEditOptions {