+++ projects/angular-cesium/src/lib/angular-cesium-widgets/models/rectangle-editor-observable.ts
@@ -2,8 +2,7 @@ import { EditPoint } from './edit-point';
 import { EditorObservable } from './editor-observable';
 import { RectangleEditUpdate } from './rectangle-edit-update';
 import { Cartesian3 } from '../../angular-cesium/models/cartesian3';
-import { PointProps } from './polyline-edit-options';
-
+import { PointProps } from './point-edit-options';
 
 export class RectangleEditorObservable extends EditorObservable<RectangleEditUpdate> {
   setManually: (