+++ projects/angular-cesium/src/lib/angular-cesium-widgets/services.ts
@@ -1,3 +1,4 @@
+export * from './services/entity-editors/points-editor/points-editor.service';
 export * from './services/entity-editors/polygons-editor/polygons-editor.service';
 export * from './services/entity-editors/rectangles-editor/rectangles-editor.service';
 export * from './services/entity-editors/circles-editor/circles-editor.service';