+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Features
+* Add points-editor component, model, and service. `Points` are now first class components that support entity features, such as labels and enable/disable.
+
 ## 0.63
 ### Features
 * Refactor demo project and create: https://github.com/articodeltd/angular-cesium-demo